fix outputs
